결제 뭔가 되고있다...!! 하하하하하하하하하

+++ matdori/src/main/java/com/kh/matdori/dto/MenuByResDto.java
@@ -12,7 +12,7 @@ public class MenuByResDto {
     private String menuTypeName;
     private int menuNo;
     private String menuName;
-    private Float menuPrice;
+    private int menuPrice;
     private String menuContent;
     private int imageId;
 }
